+++ packages/app-store/stripepayment/api/portal.ts
@@ -1,109 +1,59 @@
 import type { NextApiRequest, NextApiResponse } from "next";
+import type { Session } from "next-auth";
 
 import { WEBAPP_URL } from "@calcom/lib/constants";
 import { getSafeRedirectUrl } from "@calcom/lib/getSafeRedirectUrl";
-import { HttpError } from "@calcom/lib/http-error";
-import logger from "@calcom/lib/logger";
-import { TeamRepository } from "@calcom/lib/server/repository/team";
-import prisma from "@calcom/prisma";
-import type { Prisma } from "@calcom/prisma/client";
-import { teamMetadataSchema } from "@calcom/prisma/zod-utils";
 
-import { getStripeCustomerIdFromUserId } from "../lib/customer";
-import stripe from "../lib/server";
-import { getSubscriptionFromId } from "../lib/subscriptions";
+import { BillingPortalServiceFactory } from "../lib/BillingPortalService";
 
-const getBillingPortalUrl = async (customerId: string, return_url: string) => {
-  const log = logger.getSubLogger({ prefix: ["getBillingPortalUrl"] });
-  try {
-    const portalSession = await stripe.billingPortal.sessions.create({
-      customer: customerId,
-      return_url,
-    });
+interface AuthenticatedUser {
+  id: number;
+}
 
-    return portalSession.url;
-  } catch (e) {
-    log.error(`Failed to create billing portal session for ${customerId}: ${e}`);
-    throw new Error("Failed to create billing portal session");
-  }
+interface RequestWithSession extends NextApiRequest {
+  session?: Session | null;
+}
+
+export const validateAuthentication = (req: NextApiRequest): AuthenticatedUser | null => {
+  const userId = (req as RequestWithSession).session?.user?.id;
+  if (!userId) return null;
+  return { id: userId };
 };
 
-const getValidatedTeamSubscriptionId = async (metadata: Prisma.JsonValue) => {
-  const teamMetadataParsed = teamMetadataSchema.safeParse(metadata);
+export const buildReturnUrl = (returnTo?: string): string => {
+  const defaultUrl = `${WEBAPP_URL}/settings/billing`;
 
-  if (!teamMetadataParsed.success) {
-    throw new HttpError({ statusCode: 400, message: "Invalid team metadata" });
-  }
+  if (typeof returnTo !== "string") return defaultUrl;
 
-  if (!teamMetadataParsed.data?.subscriptionId) {
-    throw new HttpError({ statusCode: 400, message: "Subscription Id not found for team" });
-  }
-
-  return teamMetadataParsed.data.subscriptionId;
-};
-
-const getValidatedTeamSubscriptionIdForPlatform = async (subscriptionId?: string | null) => {
-  if (!subscriptionId) {
-    throw new HttpError({ statusCode: 400, message: "Subscription Id not found for team" });
-  }
-
-  return subscriptionId;
+  const safeRedirectUrl = getSafeRedirectUrl(returnTo);
+  return safeRedirectUrl || defaultUrl;
 };
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
-  if (req.method !== "POST" && req.method !== "GET")
+  if (req.method !== "POST" && req.method !== "GET") {
     return res.status(405).json({ message: "Method not allowed" });
+  }
 
-  if (!req.session?.user?.id) return res.status(401).json({ message: "Not authenticated" });
+  const user = validateAuthentication(req);
+  if (!user) {
+    return res.status(401).json({ message: "Not authenticated" });
+  }
 
-  const userId = req.session.user.id;
   const teamId = req.query.teamId ? parseInt(req.query.teamId as string) : null;
-  let return_url = `${WEBAPP_URL}/settings/billing`;
-  if (!teamId) {
-    const customerId = await getStripeCustomerIdFromUserId(userId);
-    if (!customerId) return res.status(404).json({ message: "CustomerId not found" });
+  const returnUrl = buildReturnUrl(req.query.returnTo as string);
 
-    const billingPortalUrl = await getBillingPortalUrl(customerId, return_url);
+  try {
+    if (!teamId) {
+      const userService = BillingPortalServiceFactory.createUserService();
+      return await userService.processBillingPortal(user.id, returnUrl, res);
+    }
 
-    return res.redirect(302, billingPortalUrl);
+    const billingService = await BillingPortalServiceFactory.createService(teamId);
+    return await billingService.processBillingPortal(user.id, teamId, returnUrl, res);
+  } catch (error) {
+    if (error instanceof Error && error.message === "Team not found") {
+      return res.status(404).json({ message: "Team not found" });
+    }
+    throw error;
   }
-
-  const teamRepository = new TeamRepository(prisma);
-  const team = await teamRepository.getTeamByIdIfUserIsAdmin({
-    teamId,
-    userId,
-  });
-
-  if (!team) return res.status(404).json({ message: "Team not found" });
-
-  if (typeof req.query.returnTo === "string") {
-    const safeRedirectUrl = getSafeRedirectUrl(req.query.returnTo);
-    if (safeRedirectUrl) return_url = safeRedirectUrl;
-  }
-
-  let subscriptionId = "";
-
-  if (team.isPlatform) {
-    subscriptionId = await getValidatedTeamSubscriptionIdForPlatform(team.platformBilling?.subscriptionId);
-  } else {
-    subscriptionId = await getValidatedTeamSubscriptionId(team.metadata);
-  }
-
-  const subscription = await getSubscriptionFromId(subscriptionId);
-
-  if (!subscription) {
-    return res.status(400).json({ message: "Subscription not found" });
-  }
-
-  if (!subscription.customer) {
-    return res.status(400).json({ message: "Subscription customer not found" });
-  }
-
-  const customerId = subscription.customer as string;
-
-  if (!customerId) return res.status(400).json({ message: "CustomerId not found in stripe" });
-
-  const billingPortalUrl = await getBillingPortalUrl(customerId, return_url);
-
-  res.redirect(302, billingPortalUrl);
 }
